requirements.txt
alembic==1.4.2
amqp==2.5.2
attrs==19.3.0
autopep8==1.5
billiard==3.6.3.0
celery==4.4.2
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
Flask==1.1.1
Flask-json-schema==0.0.5
Flask-SQLAlchemy==2.4.1
idna==2.9
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
kombu==4.6.8
Mako==1.1.2
MarkupSafe==1.1.1
numpy==1.18.2
pandas==1.0.3
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pyrsistent==0.15.7
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
pytz==2019.3
redis==3.4.1
requests==2.23.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
SQLAlchemy==1.3.15
urllib3==1.25.8
uWSGI==2.0.18
vine==1.3.0
Werkzeug==1.0.0
zipp==3.1.0
